Worked on cli
