+++ Cargo.toml
@@ -9,7 +9,9 @@ edition = "2021"
 actix = "0.13.1"
 actix-web = "4.4.0"
 anyhow = "1.0.75"
+cfg-if = "1.0.0"
 chrono = { version = "0.4.31", features = ["serde"] }
+clap = { version = "4.4.4", features = ["string", "env", "derive", "debug"] }
 httptest = "0.15.4"
 hyper = { version = "0.14.27", features = ["full"] }
 hyper-tls = "0.5.0"
@@ -23,4 +25,9 @@ tokio = { version = "1.32.0", features = ["full"] }
 uuid = { version = "1.4.1", features = ["serde", "v4"] }
 
 [profile.dev]
-opt-level = 3 
+opt-level = 3
+
+[features]
+default = ["web"]
+web = []
+cli = []
